+++ example/lib/set_or_delete_location.dart
@@ -15,7 +15,7 @@ class SetOrDeleteLocationDialog extends StatelessWidget {
 
   final String id;
   final String name;
-  final GeoLatLon geoLatLon;
+  final GeoFirePoint geoLatLon;
 
   @override
   Widget build(BuildContext context) {