+++ lib/geo_lat_lon.dart
@@ -2,6 +2,6 @@ export 'src/converter/i_converter/i_fire_store_model_converter.dart';
 export 'src/entity/geo.dart';
 export 'src/entity/location.dart';
 export 'src/geo_collection_reference.dart';
-export 'src/geo_lat_lon.dart';
+export 'src/geo_fire_point.dart';
 export 'src/utils/neighbor_geohashes.dart';
 export 'src/utils/typed_collection_reference.dart';