+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the `geo_lat_lon` package will be documented in this file
 
 ---
 
+## 0.0.3
+
+### Aug 6, 2025
+
+### ✅Refactored
+
+* ✅ Renamed `GeoLatLon` to `GeoFirePoint`.
+
 
 ## 0.0.2
 
@@ -14,7 +22,6 @@ All notable changes to the `geo_lat_lon` package will be documented in this file
 * ✅ Updated the Example code with `GoogleMapExample`.
 * ✅ Updated the [README.md](README.md) with more Example codes.
 
-
 ## 0.0.1
 
 ### Aug 5, 2025
@@ -23,7 +30,7 @@ All notable changes to the `geo_lat_lon` package will be documented in this file
 
 * ✅ Introduced `GeoCollectionReference<T>` for Firestore geo queries.
 * ✅ Added support for radius-based filtering with `fetchWithin` and `subscribeWithin`.
-* ✅ Implemented `GeoLatLon` class to wrap `GeoPoint` and generate geohashes.
+* ✅ Implemented `GeoFirePoint` class to wrap `GeoPoint` and generate geohashes.
 * ✅ Added `GeoDocumentSnapshot<T>` for distance-aware querying.
 * ✅ Provided `strictMode` for accurate client-side distance filtering.
 * ✅ Utility support for calculating geohash neighbors.